cheat_manager: improve parser errors
