+++ rpcs3/rpcs3qt/cheat_manager.h
@@ -8,32 +8,7 @@
 #include <QComboBox>
 #include <QPushButton>
 
-enum class cheat_type : u8
-{
-	unsigned_8_cheat,
-	unsigned_16_cheat,
-	unsigned_32_cheat,
-	unsigned_64_cheat,
-	signed_8_cheat,
-	signed_16_cheat,
-	signed_32_cheat,
-	signed_64_cheat,
-	max
-};
-
-constexpr u8 cheat_type_max = static_cast<u8>(cheat_type::max);
-
-struct cheat_info
-{
-	std::string game;
-	std::string description;
-	cheat_type type = cheat_type::max;
-	u32 offset{};
-	std::string red_script;
-
-	bool from_str(const std::string& cheat_line);
-	std::string to_str() const;
-};
+#include "Utilities/cheat_info.h"
 
 class cheat_engine
 {
