+++ rpcs3/util/yaml.cpp
@@ -1,5 +1,32 @@
 ﻿#include "util/yaml.hpp"
 #include "Utilities/types.h"
+#include "Utilities/cheat_info.h"
+#include "Utilities/Config.h"
+
+namespace YAML
+{
+	template <>
+	struct convert<cheat_info>
+	{
+		static bool decode(const Node& node, cheat_info& rhs)
+		{
+			if (node.size() != 3)
+			{
+				return false;
+			}
+
+			rhs.description = node[0].Scalar();
+			u64 type64      = 0;
+			if (!cfg::try_to_enum_value(&type64, &fmt_class_string<cheat_type>::format, node[1].Scalar()))
+				return false;
+			if (type64 >= cheat_type_max)
+				return false;
+			rhs.type       = cheat_type{::narrow<u8>(type64)};
+			rhs.red_script = node[2].Scalar();
+			return true;
+		}
+	};
+} // namespace YAML
 
 std::pair<YAML::Node, std::string> yaml_load(const std::string& from)
 {
@@ -32,5 +59,7 @@ T get_yaml_node_value(YAML::Node node, std::string& error_message)
 	return {};
 }
 
+template u32 get_yaml_node_value<u32>(YAML::Node, std::string&);
 template u64 get_yaml_node_value<u64>(YAML::Node, std::string&);
 template f64 get_yaml_node_value<f64>(YAML::Node, std::string&);
+template cheat_info get_yaml_node_value<cheat_info>(YAML::Node, std::string&);
